Feat (mirror-encryption): initial files
Add module and sample input and output files for the [mirror encryption
problem](https://www.reddit.com/r/dailyprogrammer/comments/4m3ddb/20160601_challenge_269_intermediate_mirror/).
The module has a main method and can be run, but also provides a library
containing a class and can thus be imported. The sample input and
outputs include the one provided in the page above, but also include a
much longer string to test the relative speed of translating each
character individually and using a precomputed table.